\name{mDINGO-package}
\alias{mDINGO-package}
\alias{mDINGO}
\docType{package}
\title{
mDINGO: Microbial Differential Network Analysis in Genomics
}
\description{
This packages jointly estimates group specific partial correlations for compositional microbiome data.
}
\details{
This packages jointly estimates group specific partial correlations.
}
\author{
Nick Berry -- originally Min Jin Ha, Veerabhadran Baladandayuthapani, and Kim-Anh Do
}
\keyword{ package }