\name{plotGraphList}
\alias{plotGraphList}

\title{
Convert igraph object into precision matrix.}
\description{
Copies graph2prec function from SpiecEase, with changes to make it work on igraph object. See that documentation (https://github.com/zdk123/SpiecEasi/blob/master/man/graph2prec.Rd) for more information.
}