\name{plotGraphList}
\alias{plotGraphList}

\title{
Plot output from gen_graphs_sf
}
\description{
Plot gen_graphs_sf graph output using same layout.
}

\arguments{
  \item{graphList}{
	List object returned from gen_graphs_sf
}
}
\value{
	2 by 2 plot matrix of igraph plots.
}
\author{
    Nick Berry berryni@iastate.edu
}
