\name{DINGO-package}
\alias{DINGO-package}
\alias{DINGO}
\docType{package}
\title{
DINGO: Differential Network Analysis in Genomics
}
\description{
This packages jointly estimates group specific partial correlations.
}
\details{
This packages jointly estimates group specific partial correlations.
}
\author{
Min Jin Ha, Veerabhadran Baladandayuthapani, and Kim-Anh Do
}
\keyword{ package }